requirements.txt
flask==3.0.2
python-dotenv==1.0.1
google-cloud-aiplatform==1.42.1
flask-cors==4.0.0
google-generativeai==0.3.2 